app/requirements.txt
acres==0.3.0
aiofiles==24.1.0
alembic==1.16.1
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
etelemetry==0.3.1
exceptiongroup==1.3.0
filelock==3.18.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frontend==0.0.3
h11==0.16.0
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
looseversion==1.3.0
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.2.1
prov==2.0.1
puremagic==1.29
pydot==4.0.0
PyMuPDF==1.26.0
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
tools==1.0.2
traits==7.0.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
